% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CrossSCC.R
\name{plot_CrossSCC}
\alias{plot_CrossSCC}
\title{Plot CrossSCC result by the interactive viewer}
\usage{
plot_CrossSCC(x)
}
\arguments{
\item{x}{CrossSCC result.}
}
\value{
a interactive plot, as a data.tree object.
}
\description{
Plot CrossSCC result by the interactive viewer
}
\examples{
data('cl.b1')
handsome.zuo <- CrossSCC(cl.b1[1:500,], ncores = 10, mean.posterior.cutoff = 0.18)
plot_CrossSCC(handsome.zuo)
}
