% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PCA.R
\name{plot_PCA}
\alias{plot_PCA}
\title{Perform PCA analysis and plot result}
\usage{
plot_PCA(x)
}
\arguments{
\item{x}{CrossSCC result.}
}
\value{
a PCA plot
}
\description{
Perform PCA analysis and plot result
}
\examples{
data('cl.b1')
handsome.zuo <- CrossSCC(cl.b1[1:500,], ncores = 10, mean.posterior.cutoff = 0.18)
plot_PCA(handsome.zuo)
}
