% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossSCC.R
\docType{package}
\name{CrossSCC}
\alias{CrossSCC}
\alias{CrossSCC-package}
\title{CrossSCC: An Single-Cell Clustering analysis framework Crossing batches of data}
\usage{
CrossSCC(m, ncores = 4, var.cutoff = 0.9, mapping = "org.Hs.eg.db",
  simple.mapping = "org.HsSimple.eg.db", mean.posterior.cutoff = 0.3,
  ovl.cutoff = 0.05, mean.posterior.weight = 0.5,
  min.group.size = 0.1, ovl.weight = 0.5, lambda.cutoff = 0.9,
  ontos = "BP", verbose = R.utils::Verbose(threshold = -1),
  show.progress.bar = TRUE)
}
\arguments{
\item{m}{a matrix of single cell expression values.}

\item{ncores}{number of CPU cores used. Caution: By default, CrossSCC will try use all CPU cores, which may cause "core dumped" error.}

\item{var.cutoff}{cutoff value when filtering features.}

\item{mapping}{database package name according to organism. e.g. "org.Hs.eg.db".}

\item{mean.posterior.cutoff}{cutoff value for mean posterior of each component when rating Gaussion Mixture Model.}

\item{ovl.cutoff}{cutoff value for OVL (overlapping coefficient) when rating Gaussion Mixture Model.}

\item{mean.posterior.weight}{the weight for mean.posterior.cutoff when assessing the performance of the model.}

\item{min.group.size}{minimal final group size (sample number). Only groups has more samples than this threshold will be further splitted.}

\item{ovl.weight}{the weight for ovl.cutoff when assessing the performance of the model.}

\item{lambda.cutoff}{Gaussian components with lambda over this cutoff value will be classified as a model representing a subtype of samples.}

\item{ontos}{ontology terms used when converting features to meta-gene features. Default is "BP", could be arbitrary combination of c('BP', 'MF', 'CC').}

\item{verbose}{verbose level. By default, CrossSCC will output all logs as well as progress bars.}

\item{show.progress.bar}{Set to FALSE if you don't want to see progress bar.}
}
\value{
a data.tree object.
}
\description{
CrossSCC: An Single-Cell Clustering analysis framework Crossing batches of data

CrossSCC
}
\examples{
data('cl.b1')
handsome.zuo <- CrossSCC(cl.b1[1:500,], ncores = 10, mean.posterior.cutoff = 0.18)
}
